+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## [0.0.2] - added footer layout
+flutter layout now has built in sticky footer layout support
+
 ## [0.0.1] - pub package registration
 this package does not have any functionality yet.